+++ src/components/Products/Categories/Categories.js
@@ -1,28 +1,24 @@
-// Categories.js
 import React, { useState, useEffect } from 'react';
-import { useHistory } from 'react-router-dom';
+import { useHistory } from 'react-router-dom/index';
 import { Table, Button, Form, Pagination, Modal } from 'react-bootstrap';
 import axios from 'axios';
-
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faPlus, faTrash, faEdit, faList } from '@fortawesome/free-solid-svg-icons';
+import { faPlus, faTrash, faEdit, faList, faArrowLeft } from '@fortawesome/free-solid-svg-icons';
 import 'bootstrap/dist/css/bootstrap.min.css';
+import Swal from 'sweetalert2';
+import '.././styles/Crud.css'
 
 const Categories = () => {
   const [categories, setCategories] = useState([]);
   const [newCategoryName, setNewCategoryName] = useState('');
-
   const [showCreateModal, setShowCreateModal] = useState(false);
   const [showUpdateModal, setShowUpdateModal] = useState(false);
   const [updateCategoryId, setUpdateCategoryId] = useState(null);
-
   const [currentPage, setCurrentPage] = useState(1);
   const [categoriesPerPage] = useState(10);
-
   const indexOfLastCategory = currentPage * categoriesPerPage;
   const indexOfFirstCategory = indexOfLastCategory - categoriesPerPage;
   const currentCategories = categories.slice(indexOfFirstCategory, indexOfLastCategory);
-
   const history = useHistory();
 
   useEffect(() => {
@@ -43,49 +39,90 @@ const Categories = () => {
       const response = await axios.post('https://localhost:7028/api/categories', {
         name: newCategoryName,
       });
-
+  
       setCategories([response.data, ...categories]);
-      
       setNewCategoryName('');
-      
       setShowCreateModal(false);
+  
+      Swal.fire(
+        '¡Éxito!',
+        '¡La categoría ha sido creada exitosamente.',
+        'success'
+      );
     } catch (error) {
       console.error('Error creating category:', error);
+  
+      Swal.fire(
+        'Error',
+        'Hubo un problema al crear la categoría.',
+        'error'
+      );
     }
   };
 
   const handleDeleteCategory = async (categoryId) => {
-    try {
-      await axios.delete(`https://localhost:7028/api/categories/${categoryId}`);
-      
-      const updatedCategories = categories.filter((category) => category.categoryId !== categoryId);
-      setCategories(updatedCategories);
-    } catch (error) {
-      console.error('Error deleting category:', error);
-    }
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: '¡No podrás revertir esto!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminarlo!'
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          await axios.delete(`https://localhost:7028/api/categories/${categoryId}`);
+
+          const updatedCategories = categories.filter((category) => category.categoryId !== categoryId);
+          
+          setCategories(updatedCategories);
+          
+          Swal.fire(
+            '¡Eliminado!',
+            '¡Tu categoría ha sido eliminada.',
+            'success'
+          );
+        } catch (error) {
+          console.error('Error deleting category:', error);
+        
+          Swal.fire(
+            'Error',
+            'Hubo un problema al eliminar la categoría.',
+            'error'
+          );
+        }
+      }
+    });
   };
 
   const handleUpdateCategory = async (categoryId, newName) => {
     try {
-        await axios.put(
-            `https://localhost:7028/api/categories/${categoryId}`, {name: newName });
-
-        // Realiza una nueva solicitud para obtener la lista actualizada
-        const response = await axios.get('https://localhost:7028/api/categories');
-
-        const updatedCategories = response.data;
-
-        // Actualiza el estado con la nueva lista
-        setCategories(updatedCategories);
-
-        setNewCategoryName({
-            Name: '',
-        });
-
-        setShowUpdateModal(false);
-
+      await axios.put(
+        `https://localhost:7028/api/categories/${categoryId}`, 
+        { name: newName }
+      );
+  
+      const response = await axios.get('https://localhost:7028/api/categories');
+      const updatedCategories = response.data;
+  
+      setCategories(updatedCategories);
+      setNewCategoryName({ Name: '' });
+      setShowUpdateModal(false);
+  
+      Swal.fire(
+        '¡Éxito!',
+        '¡La categoría ha sido actualizada exitosamente.',
+        'success'
+      );
     } catch (error) {
-        console.error('Error updating rawMaterial:', error);
+      console.error('Error updating category:', error);
+  
+      Swal.fire(
+        'Error',
+        'Hubo un problema al actualizar la categoría.',
+        'error'
+      );
     }
   };
 
@@ -101,26 +138,19 @@ const Categories = () => {
     history.goBack();
   };
 
-
   return (
     <div className="categories-container">
-
-      <br/>
-
-      <h2>Modulo Categorías</h2>
-
       <br />
-
-      <Button variant="primary" onClick={() => setShowCreateModal(true)}>
-        <FontAwesomeIcon icon={faPlus} /> Nueva Categoría
-      </Button>
-
-      <Button variant="danger" onClick={handleGoBack}>
-        Volver
-      </Button>
-
-      <hr/>
-
+      <h2>Modulo Categorías</h2>
+      <div className="crud-buttons">
+        <Button variant="primary" onClick={() => setShowCreateModal(true)} style={{ marginRight: '10px' }}>
+          <FontAwesomeIcon icon={faPlus} /> Nueva Categoría
+        </Button>
+        <Button variant="danger" onClick={handleGoBack} style={{ marginRight: '10px' }}> 
+          <FontAwesomeIcon icon={faArrowLeft} /> Volver
+        </Button>
+      </div>
+      <hr />
       <Table striped bordered hover>
         <thead>
           <tr>
@@ -133,25 +163,34 @@ const Categories = () => {
             <tr key={category.categoryId}>
               <td>{category.name}</td>
               <td>
-                <Button variant="info" onClick={() => {
-                  setUpdateCategoryId(category.categoryId);
-                  setShowUpdateModal(true);
-                }}>
-                  <FontAwesomeIcon icon={faEdit} /> Actualizar
-                </Button>{' '}
-                <Button variant="danger" onClick={() => handleDeleteCategory(category.categoryId)}>
-                  <FontAwesomeIcon icon={faTrash} /> Eliminar
-                </Button>
-                <Button variant="primary" onClick={() => handleShowSubcategories(category.categoryId)}>
-                  <FontAwesomeIcon icon={faList} /> Ver Subcategorías
-                </Button>
+                <div className="action-button">
+                  <Button
+                    variant="primary"
+                    onClick={() => {
+                      setUpdateCategoryId(category.categoryId);
+                      setShowUpdateModal(true);
+                    }}
+                  >
+                    <FontAwesomeIcon icon={faEdit} /> Actualizar
+                  </Button>
+                  <Button
+                    variant="danger"
+                    onClick={() => handleDeleteCategory(category.categoryId)}
+                  >
+                    <FontAwesomeIcon icon={faTrash} /> Eliminar
+                  </Button>
+                  <Button
+                    variant="info"
+                    onClick={() => handleShowSubcategories(category.categoryId)}
+                  >
+                    <FontAwesomeIcon icon={faList} /> Ver Subcategorías
+                  </Button>
+                </div>
               </td>
             </tr>
           ))}
         </tbody>
       </Table>
-
-      {/* Componente de paginación */}
       <Pagination>
         {Array.from({ length: Math.ceil(categories.length / categoriesPerPage) }).map((_, index) => (
           <Pagination.Item
@@ -163,8 +202,6 @@ const Categories = () => {
           </Pagination.Item>
         ))}
       </Pagination>
-
-      {/* Modal para crear categoría */}
       <Modal show={showCreateModal} onHide={() => setShowCreateModal(false)}>
         <Modal.Header closeButton>
           <Modal.Title>Nueva Categoría</Modal.Title>
@@ -172,7 +209,7 @@ const Categories = () => {
         <Modal.Body>
           <Form>
             <Form.Group controlId="formCategoryName">
-              <Form.Label>Nombre</Form.Label>
+              <Form.Label><b>Nombre</b></Form.Label>
               <Form.Control
                 type="text"
                 placeholder="Ingresa el nombre de la categoría"
@@ -191,8 +228,6 @@ const Categories = () => {
           </Button>
         </Modal.Footer>
       </Modal>
-
-      {/* Modal para actualizar categoría */}
       <Modal show={showUpdateModal} onHide={() => setShowUpdateModal(false)}>
         <Modal.Header closeButton>
           <Modal.Title>Actualizar Categoría</Modal.Title>
@@ -200,7 +235,7 @@ const Categories = () => {
         <Modal.Body>
           <Form>
             <Form.Group controlId="formUpdatedCategoryName">
-              <Form.Label>Nombre</Form.Label>
+              <Form.Label><b>Nombre</b></Form.Label>
               <Form.Control
                 type="text"
                 placeholder="Ingresa el nuevo nombre de la categoría"
